Fix a warning.

+++ crawl-ref/source/food.cc
@@ -725,7 +725,7 @@ bool prompt_eat_inventory_item(int slot)
                 prompt_invent_item(you.species == SP_VAMPIRE ? "Drain what?"
                                                              : "Eat which item?",
                                    MT_INVLIST,
-                                   you.species == SP_VAMPIRE ? OSEL_VAMP_EAT
+                                   you.species == SP_VAMPIRE ? (int)OSEL_VAMP_EAT
                                                              : OBJ_FOOD,
                                    true, true, true, 0, -1, NULL,
                                    OPER_EAT);
